refactor: moved extra stuff away from rusty-dice
