+++ src/main.rs
@@ -3,6 +3,7 @@ use app::App;
 
 mod app;
 mod cli;
+mod dice;
 mod expressions;
 
 fn main() -> Result<()> {
